rename useNftRetry
