+++ src/scenes/NftDetailPage/NftDetailImage.tsx
@@ -9,7 +9,7 @@ import { useMemo } from 'react';
 import { StyledVideo } from './NftDetailVideo';
 import noop from 'utils/noop';
 import { CouldNotRenderNftError } from 'errors/CouldNotRenderNftError';
-import { useThrowOnMediaFailure } from 'hooks/useNftDisplayRetryLoader';
+import { useThrowOnMediaFailure } from 'hooks/useNftRetry';
 
 type Props = {
   tokenRef: NftDetailImageFragment$key;
